requirements.txt
# Core
numpy>=1.24.0
h5py>=3.8.0
pyyaml>=6.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0

pytest-benchmark>=4.0.0
locust>=2.20.0
memory-profiler>=0.61.0
line-profiler>=4.1.0
scipy>=1.11.0  # For generating realistic random load data

prometheus-client>=0.17.0
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-instrumentation-fastapi>=0.40b0
